// web/src/openapi/models/TracingRecordScene.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tracing-lv-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const TracingRecordScene = {
    Tree: 'Tree',
    SpanField: 'SpanField',
    SpanEnter: 'SpanEnter'
} as const;
export type TracingRecordScene = typeof TracingRecordScene[keyof typeof TracingRecordScene];


export function instanceOfTracingRecordScene(value: any): boolean {
    for (const key in TracingRecordScene) {
        if (Object.prototype.hasOwnProperty.call(TracingRecordScene, key)) {
            if (TracingRecordScene[key as keyof typeof TracingRecordScene] === value) {
                return true;
            }
        }
    }
    return false;
}

export function TracingRecordSceneFromJSON(json: any): TracingRecordScene {
    return TracingRecordSceneFromJSONTyped(json, false);
}

export function TracingRecordSceneFromJSONTyped(json: any, ignoreDiscriminator: boolean): TracingRecordScene {
    return json as TracingRecordScene;
}

export function TracingRecordSceneToJSON(value?: TracingRecordScene | null): any {
    return value as any;
}

export function TracingRecordSceneToJSONTyped(value: any, ignoreDiscriminator: boolean): TracingRecordScene {
    return value as TracingRecordScene;
}