// web/src/openapi/models/TracingTreeRecordDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tracing-lv-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TracingRecordDto } from './TracingRecordDto';
import {
    TracingRecordDtoFromJSON,
    TracingRecordDtoFromJSONTyped,
    TracingRecordDtoToJSON,
    TracingRecordDtoToJSONTyped,
} from './TracingRecordDto';
import type { TracingTreeEndDto } from './TracingTreeEndDto';
import {
    TracingTreeEndDtoFromJSON,
    TracingTreeEndDtoFromJSONTyped,
    TracingTreeEndDtoToJSON,
    TracingTreeEndDtoToJSONTyped,
} from './TracingTreeEndDto';
import type { TracingTreeRecordVariantDto } from './TracingTreeRecordVariantDto';
import {
    TracingTreeRecordVariantDtoFromJSON,
    TracingTreeRecordVariantDtoFromJSONTyped,
    TracingTreeRecordVariantDtoToJSON,
    TracingTreeRecordVariantDtoToJSONTyped,
} from './TracingTreeRecordVariantDto';

/**
 * 
 * @export
 * @interface TracingTreeRecordDto
 */
export interface TracingTreeRecordDto {
    /**
     * 
     * @type {TracingTreeEndDto}
     * @memberof TracingTreeRecordDto
     */
    end?: TracingTreeEndDto | null;
    /**
     * 
     * @type {TracingRecordDto}
     * @memberof TracingTreeRecordDto
     */
    record: TracingRecordDto;
    /**
     * 
     * @type {TracingTreeRecordVariantDto}
     * @memberof TracingTreeRecordDto
     */
    variant?: TracingTreeRecordVariantDto | null;
}

/**
 * Check if a given object implements the TracingTreeRecordDto interface.
 */
export function instanceOfTracingTreeRecordDto(value: object): value is TracingTreeRecordDto {
    if (!('record' in value) || value['record'] === undefined) return false;
    return true;
}

export function TracingTreeRecordDtoFromJSON(json: any): TracingTreeRecordDto {
    return TracingTreeRecordDtoFromJSONTyped(json, false);
}

export function TracingTreeRecordDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TracingTreeRecordDto {
    if (json == null) {
        return json;
    }
    return {
        
        'end': json['end'] == null ? undefined : TracingTreeEndDtoFromJSON(json['end']),
        'record': TracingRecordDtoFromJSON(json['record']),
        'variant': json['variant'] == null ? undefined : TracingTreeRecordVariantDtoFromJSON(json['variant']),
    };
}

  export function TracingTreeRecordDtoToJSON(json: any): TracingTreeRecordDto {
      return TracingTreeRecordDtoToJSONTyped(json, false);
  }

  export function TracingTreeRecordDtoToJSONTyped(value?: TracingTreeRecordDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'end': TracingTreeEndDtoToJSON(value['end']),
        'record': TracingRecordDtoToJSON(value['record']),
        'variant': TracingTreeRecordVariantDtoToJSON(value['variant']),
    };
}