// web/src/openapi/models/TracingTreeRecordVariantDtoOneOf.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tracing-lv-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TracingSpanRunDto } from './TracingSpanRunDto';
import {
    TracingSpanRunDtoFromJSON,
    TracingSpanRunDtoFromJSONTyped,
    TracingSpanRunDtoToJSON,
    TracingSpanRunDtoToJSONTyped,
} from './TracingSpanRunDto';

/**
 * 
 * @export
 * @interface TracingTreeRecordVariantDtoOneOf
 */
export interface TracingTreeRecordVariantDtoOneOf {
    /**
     * 
     * @type {TracingSpanRunDto}
     * @memberof TracingTreeRecordVariantDtoOneOf
     */
    spanRun: TracingSpanRunDto;
}

/**
 * Check if a given object implements the TracingTreeRecordVariantDtoOneOf interface.
 */
export function instanceOfTracingTreeRecordVariantDtoOneOf(value: object): value is TracingTreeRecordVariantDtoOneOf {
    if (!('spanRun' in value) || value['spanRun'] === undefined) return false;
    return true;
}

export function TracingTreeRecordVariantDtoOneOfFromJSON(json: any): TracingTreeRecordVariantDtoOneOf {
    return TracingTreeRecordVariantDtoOneOfFromJSONTyped(json, false);
}

export function TracingTreeRecordVariantDtoOneOfFromJSONTyped(json: any, ignoreDiscriminator: boolean): TracingTreeRecordVariantDtoOneOf {
    if (json == null) {
        return json;
    }
    return {
        
        'spanRun': TracingSpanRunDtoFromJSON(json['spanRun']),
    };
}

  export function TracingTreeRecordVariantDtoOneOfToJSON(json: any): TracingTreeRecordVariantDtoOneOf {
      return TracingTreeRecordVariantDtoOneOfToJSONTyped(json, false);
  }

  export function TracingTreeRecordVariantDtoOneOfToJSONTyped(value?: TracingTreeRecordVariantDtoOneOf | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'spanRun': TracingSpanRunDtoToJSON(value['spanRun']),
    };
}